[T1.2]Read data to graph done! :D

+++ DeliverySite.cpp
@@ -6,5 +6,5 @@
 
 #include <utility>
 
-DeliverySite::DeliverySite(std::string  city, int id, const std::string& code, double demand, int population)
+DeliverySite::DeliverySite(std::string  city, int id, const std::string& code, double demand, double population)
         : city(std::move(city)), GraphNode(id, code), demand(demand), population(population) {}
